GraphicEngine and many other related stuff fixed
Still requires some works

+++ src/Graphics/UI/UI_element.h
@@ -5,8 +5,8 @@ class UI_element
 
 protected:
 	UI_element(float x, float y) :x(x), y(y) {};
-	int height;
-	int width;
+	float height;
+	float width;
 
 	float x;
 	float y;
@@ -15,6 +15,6 @@ public:
 	float GetX() const { return x; }
 	float GetY() const { return y; }
 
-	int GetHeight() const { return height; }
-	int GetWidth() const { return width; }
+	float GetHeight() const { return height; }
+	float GetWidth() const { return width; }
 };
